move Pegasus markdown renderer into new 'common' filepath
