+++ lib/cdo/pegasus/actionview_sinatra.rb
@@ -1,5 +1,6 @@
 require 'action_view'
 require 'cdo/markdown_handler'
+require 'cdo/markdown/renderer'
 require 'cdo/redcarpet/inline'
 require 'haml'
 require 'haml/template'
@@ -7,46 +8,12 @@ require 'redcarpet'
 require 'sinatra/base'
 
 module ActionViewSinatra
-  class MarkdownRenderer < Redcarpet::Render::HTML
-    OPTIONS = {
-      autolink: true,
-      tables: true,
-      space_after_headers: true,
-      fenced_code_blocks: true,
-      lax_spacing: true
-    }
-
-    def postprocess(full_document)
-      process_div_brackets(full_document)
-    end
-
-    private
-
-    # CDO-Markdown div_brackets extension.
-    # Convert `[tag]...[/tag]` to `<div class='tag'>...</div>`.
-    def process_div_brackets(full_document)
-      full_document.
-        gsub(/<p>\[\/(.*)\]<\/p>/, '</div>').
-        gsub(/<p>\[(.*)\]<\/p>/) do
-        value = $1
-        if value[0] == '#'
-          attribute = 'id'
-          value = value[1..-1]
-        else
-          attribute = 'class'
-        end
-
-        "<div #{attribute}='#{value}'>"
-      end
-    end
-  end
-
   class Base < ActionView::Base
     def initialize(sinatra, *args)
       @sinatra = sinatra
       super(*args)
 
-      MarkdownHandler.register(MarkdownRenderer, MarkdownRenderer::OPTIONS)
+      MarkdownHandler.register(Cdo::Markdown::Renderer, Cdo::Markdown::Renderer::OPTIONS)
       MarkdownHandler.register_html_safe(Redcarpet::Render::Safe)
       MarkdownHandler.register_inline(Redcarpet::Render::Inline.new({filter_html: true}))
     end
